pyproject.toml
[project]
name = "amfk-probafeladat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pytest>=9.0.2",
    "pytest-html>=4.2.0",
    "pytest-playwright>=0.7.2",
]
